+++ common/src/commonMain/kotlin/com/projeto/nutrilog/domain/usecase/ClearUserUseCase.kt
@@ -1,0 +1,9 @@
+package com.projeto.nutrilog.domain.usecase
+
+import com.projeto.nutrilog.domain.repository.UserRepository
+
+class ClearUserUseCase(private val repository: UserRepository) {
+    suspend operator fun invoke() {
+        repository.clearUser()
+    }
+}