+++ common/src/commonMain/kotlin/com/projeto/nutrilog/domain/usecase/GetUserUseCase.kt
@@ -1,0 +1,10 @@
+package com.projeto.nutrilog.domain.usecase
+
+import com.projeto.nutrilog.database.UserEntity
+import com.projeto.nutrilog.domain.repository.UserRepository
+
+class GetUserUseCase(private val repository: UserRepository) {
+    suspend operator fun invoke(): UserEntity? {
+        return repository.getUser()
+    }
+}